refactor: extract api model
